refactor: finish moving things to components pkg
