+++ src/lib/ui/components/clicker/ClickerFooter.svelte
