+++ src/lib/ui/components/clicker/ClickerGame.svelte
