+++ src/lib/ui/components/clicker/ClickerHeader.svelte
