+++ src/lib/ui/components/clicker/ShopItem.svelte
