+++ src/lib/ui/components/clicker/cursor/ChopperCursor.svelte
