+++ src/lib/ui/components/clicker/enemies.ts
