+++ src/lib/ui/components/clicker/game.svelte.ts
