+++ src/lib/ui/components/clicker/hotbar/ClickerFace.svelte
