+++ src/lib/ui/components/clicker/hotbar/ClickerHotbar.svelte
