+++ src/lib/ui/components/clicker/shop.ts
