+++ src/lib/ui/components/clicker/windows/GameFactory.svelte
