+++ src/lib/ui/components/clicker/windows/GameInfo.svelte
