+++ src/lib/ui/components/clicker/windows/GameInfoContainer.svelte
