+++ src/lib/ui/components/clicker/windows/GameInfoIcon.svelte
