+++ src/lib/ui/components/clicker/windows/GameInventory.svelte
