+++ src/lib/ui/components/clicker/windows/GameLeaderboard.svelte
