+++ src/lib/ui/components/clicker/windows/GameSettings.svelte
