+++ src/lib/ui/components/clicker/windows/GameShop.svelte
