+++ src/lib/ui/components/clicker/windows/GameStatistics.svelte
