+++ src/lib/ui/components/clicker/windows/GameWardrobe.svelte
