+++ src/lib/ui/components/clicker/windows/GameWindow.svelte
