+++ src/lib/ui/components/clicker/windows/chop/DamageOverlay.svelte
