+++ src/lib/ui/components/clicker/windows/chop/GameChop.svelte
