+++ src/lib/ui/components/clicker/windows/chop/overlay/destroy_stage_0.png
