+++ src/lib/ui/components/clicker/windows/chop/overlay/destroy_stage_2.png
