+++ src/lib/ui/components/clicker/windows/chop/overlay/destroy_stage_3.png
