+++ src/lib/ui/components/clicker/windows/chop/overlay/destroy_stage_4.png
