+++ src/lib/ui/components/clicker/windows/chop/overlay/empty.png
