+++ src/lib/ui/components/clicker/windows/windows.svelte.ts
